Memory tuning (#82)
* Enforce Copy-on-write semantics for core repo data structures & load large data files from memory mapped files

* dotnet format

* Refactoring + Slight memory-tuning:

* Use global alias for ScanRootId, FileId, and DirId
* Change type of ScanRootId, FileId, and DirId from Int64 to Int32 (~21%
  reduction in memory and aggregate size of repo)

* dotnet format

* Use custome IBufferWriter impl for serialising SegmentedMap

+++ MemPackBench/FileDirIndexStateBench/DirHandle.cs
@@ -3,7 +3,7 @@ using MemoryPack;
 namespace MemPackBench.FileDirIndexStateBench;
 
 [MemoryPackable(SerializeLayout.Sequential)]
-public readonly partial record struct DirHandle(long ScanRootId = -1, int Index = -1)
+public readonly partial record struct DirHandle(int ScanRootId = -1, int Index = -1)
 {
     public bool IsValid => Index >= 0;
 }
