+++ MemPackBench/FileDirIndexStateBench/FileHandle.cs
@@ -3,7 +3,7 @@ using MemoryPack;
 namespace MemPackBench.FileDirIndexStateBench;
 
 [MemoryPackable(SerializeLayout.Sequential)]
-public readonly partial record struct FileHandle(long ScanRootId = -1, int Index = -1)
+public readonly partial record struct FileHandle(int ScanRootId = -1, int Index = -1)
 {
     public bool IsValid => Index >= 0;
 }
